Fix structure of the file
